# docs/openapi.json
{
 "components": {
  "schemas": {
...
   "DeviceCreated": {
    "properties": {
     "active": {
      "title": "Active",
      "type": "boolean"
     },
     "device_id": {
      "title": "Device Id",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "key": {
      "title": "Key",
      "type": "string"
     },
     "last_seen": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Seen"
     },
     "site_id": {
      "title": "Site Id",
      "type": "string"
     }
    },
    "required": [
     "id",
     "device_id",
     "site_id",
     "active",
     "last_seen",
     "key"
    ],
    "title": "DeviceCreated",
    "type": "object"
   },
   "DeviceIn": {
    "properties": {
     "device_id": {
      "title": "Device Id",
      "type": "string"
     },
     "site_id": {
      "title": "Site Id",
      "type": "string"
     }
    },
    "required": [
     "device_id",
     "site_id"
    ],
    "title": "DeviceIn",
    "type": "object"
   },
   "DeviceOut": {
    "properties": {
     "active": {
      "title": "Active",
      "type": "boolean"
     },
     "device_id": {
      "title": "Device Id",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "last_seen": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Seen"
     },
     "site_id": {
      "title": "Site Id",
      "type": "string"
     }
    },
    "required": [
     "id",
     "device_id",
     "site_id",
     "active",
     "last_seen"
    ],
    "title": "DeviceOut",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "RebuildIn": {
    "properties": {
     "confirm_site_id": {
      "title": "Confirm Site Id",
      "type": "string"
     },
     "site_id": {
      "title": "Site Id",
      "type": "string"
     }
    },
    "required": [
     "site_id",
     "confirm_site_id"
    ],
    "title": "RebuildIn",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "OceanKind",
  "version": "2.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/admin/devices": {
   "get": {
    "operationId": "list_devices_api_admin_devices_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/DeviceOut"
         },
         "title": "Response List Devices Api Admin Devices Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Devices",
    "tags": [
     "admin"
    ]
   },
   "post": {
    "operationId": "create_device_api_admin_devices_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DeviceIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceCreated"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Device",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/admin/devices/{device_pk}": {
   "delete": {
    "description": "Revocation. The unit gets 401 on its next poll and keeps its last valid\nconfig, per DATA-CONTRACT.md expiry semantics.",
    "operationId": "delete_device_api_admin_devices__device_pk__delete",
    "parameters": [
     {
      "in": "path",
      "name": "device_pk",
      "required": true,
      "schema": {
       "title": "Device Pk",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Device",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/admin/devices/{device_pk}/config": {
   "get": {
    "operationId": "get_device_config_api_admin_devices__device_pk__config_get",
    "parameters": [
     {
      "in": "path",
      "name": "device_pk",
      "required": true,
      "schema": {
       "title": "Device Pk",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceConfigOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Device Config",
    "tags": [
     "admin"
    ]
   },
   "put": {
    "description": "Tune, then publish.\n\nFull replace. Missing fields take defaults; unknown fields are an error,\nbecause a typo'd key that silently tuned nothing is a quiet failure.\n\nThe saved values are then written to `sites/{site_id}/remote_config.json`,\nsigned. That blob is the delivery path (D-020) \u2014 saving without publishing\nwould leave the panel showing a configuration no device will ever apply.",
    "operationId": "put_device_config_api_admin_devices__device_pk__config_put",
    "parameters": [
     {
      "in": "path",
      "name": "device_pk",
      "required": true,
      "schema": {
       "title": "Device Pk",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "additionalProperties": true,
        "title": "Body",
        "type": "object"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceConfigOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Device Config",
    "tags": [
     "admin"
    ]
   }
  },
...
  "/api/admin/index/rebuild": {
   "post": {
    "description": "Drop this site's index rows and rebuild them from object storage (R-12.2).\n\n**This is the proof that the index is derived.** If a rebuild cannot\nreproduce it, then something existed only in Postgres and the index had\nquietly become a second source of truth \u2014 the exact thing D-021 rules out.\nWorth running deliberately every so often, not only when something looks\nwrong.\n\nLosing these rows costs a pass, never a detection: object storage is the\nrecord and every row here came from a blob that is still there.",
    "operationId": "rebuild_index_api_admin_index_rebuild_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RebuildIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReconcileOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Rebuild Index",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/admin/index/reconcile": {
   "post": {
    "description": "Run the pass now, for one site or all of them.\n\nSafe to press twice: every write goes through the indexer, which is\nidempotent on `event_id` (R-12.3). It does not replace the timer \u2014 it exists\nso somebody investigating drift can act on it without waiting a day.",
    "operationId": "run_reconcile_api_admin_index_reconcile_post",
    "parameters": [
     {
      "in": "query",
      "name": "site_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Site Id"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ReconcileOut"
         },
         "title": "Response Run Reconcile Api Admin Index Reconcile Post",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Run Reconcile",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/admin/sites": {
   "get": {
    "operationId": "list_sites_admin_api_admin_sites_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/app__routers__admin__SitesOut"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Sites Admin",
    "tags": [
     "admin"
    ]
   },
   "post": {
    "operationId": "create_site_api_admin_sites_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Create Site Api Admin Sites Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Site",
    "tags": [
     "admin"
    ]
   }
  },
...
  "/api/admin/sites/{site_id}": {
   "delete": {
    "description": "Refuses while anything still points at it. Deleting a site out from under\na device would leave a credential valid for a site that does not exist.",
    "operationId": "delete_site_api_admin_sites__site_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Site",
    "tags": [
     "admin"
    ]
   },
   "put": {
    "description": "`site_id` itself is immutable: it is a storage path segment, and renaming\nit would orphan every event, clip and rollup already written under it.",
    "operationId": "update_site_api_admin_sites__site_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Update Site Api Admin Sites  Site Id  Put",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Site",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/admin/users": {
   "get": {
    "operationId": "list_users_api_admin_users_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/UserOut"
         },
         "title": "Response List Users Api Admin Users Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Users",
    "tags": [
     "admin"
    ]
   },
   "post": {
    "operationId": "create_user_api_admin_users_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create User",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/admin/users/{user_id}": {
   "delete": {
    "operationId": "delete_user_api_admin_users__user_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "title": "User Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete User",
    "tags": [
     "admin"
    ]
   }
  },
...
 }
}